Catch StopIteration exception when reading csv file
When the Emotiv class uses a csv file as the input source, the
setup_reader loop eventually encounters a StopIteration exception
when the csv reader reaches the end of the file.  This was not
handled, so the exception is now caught.  When the exception is
encountered, the loop finishes and Emotiv.running is set to False.